feat(balance): credit as decimal

+++ src/app/usecases/balance.py
@@ -1,3 +1,5 @@
+from decimal import Decimal
+
 from sqlalchemy.orm import Session
 
 from infra.db.models.balance import Balance
@@ -9,7 +11,7 @@ class BalanceUseCase:
         self.session = session
         self._repo = BalanceRepositry(session)
 
-    def topup_user_credits(self, user_id: int, credit_amount: int) -> Balance:
+    def topup_user_credits(self, user_id: int, credit_amount: Decimal) -> Balance:
         with self.session.begin():
             balance = self._repo._topup_user_credits(user_id=user_id, amount=credit_amount)
             return balance
